+++ src/components/Hero/Hero.test.tsx
@@ -3,7 +3,6 @@ import { render, screen } from "@testing-library/react";
 import { axe } from "jest-axe";
 import renderer from "react-test-renderer";
 
-import Image from "../Image/Image";
 import Heading from "../Heading/Heading";
 import { HeadingLevels } from "../Heading/HeadingTypes";
 import { HeroTypes } from "./HeroTypes";
@@ -19,9 +18,8 @@ export const subHeaderText = (
 );
 export const otherSubHeaderText =
   "With 92 locations across the Bronx, Manhattan, and Staten Island, The New York Public Library is an essential part of neighborhoods across the city. Visit us today.";
-export const imageComponent = (
-  <Image src="https://placeimg.com/800/400/animals" alt="Image example" />
-);
+const imageAlt = "Image example";
+const imageSrc = "https://placeimg.com/800/400/animals";
 
 describe("Hero", () => {
   describe("axe accessbility tests", () => {
@@ -54,8 +52,9 @@ describe("Hero", () => {
               text="Hero Secondary"
             />
           }
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={subHeaderText}
-          imageComponent={imageComponent}
         />
       );
       expect(await axe(container)).toHaveNoViolations();
@@ -81,6 +80,7 @@ describe("Hero", () => {
     it("passes for type Campaign", async () => {
       const { container } = render(
         <Hero
+          backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
           heroType={HeroTypes.Campaign}
           heading={
             <Heading
@@ -89,9 +89,9 @@ describe("Hero", () => {
               text="Hero Campaign"
             />
           }
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={otherSubHeaderText}
-          backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
-          imageComponent={imageComponent}
         />
       );
       expect(await axe(container)).toHaveNoViolations();
@@ -101,8 +101,9 @@ describe("Hero", () => {
       const { container } = render(
         <Hero
           heroType={HeroTypes.FiftyFifty}
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={otherSubHeaderText}
-          imageComponent={imageComponent}
         />
       );
       expect(await axe(container)).toHaveNoViolations();
@@ -144,8 +145,9 @@ describe("Hero", () => {
             text="Hero Secondary"
           />
         }
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
         subHeaderText={subHeaderText}
-        imageComponent={imageComponent}
       />
     );
 
@@ -181,6 +183,7 @@ describe("Hero", () => {
   it("renders Campaign Hero", () => {
     render(
       <Hero
+        backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
         heroType={HeroTypes.Campaign}
         heading={
           <Heading
@@ -189,9 +192,9 @@ describe("Hero", () => {
             text="Hero Campaign"
           />
         }
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
         subHeaderText={otherSubHeaderText}
-        backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
-        imageComponent={imageComponent}
       />
     );
 
@@ -212,8 +215,9 @@ describe("Hero", () => {
     render(
       <Hero
         heroType={HeroTypes.FiftyFifty}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
         subHeaderText={otherSubHeaderText}
-        imageComponent={imageComponent}
       />
     );
 
@@ -249,11 +253,12 @@ describe("Hero", () => {
 
     rerender(
       <Hero
-        heroType={HeroTypes.FiftyFifty}
-        subHeaderText={otherSubHeaderText}
-        imageComponent={imageComponent}
-        foregroundColor="#123456"
         backgroundColor="#654321"
+        foregroundColor="#123456"
+        heroType={HeroTypes.FiftyFifty}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
+        subHeaderText={otherSubHeaderText}
       />
     );
 
@@ -264,6 +269,8 @@ describe("Hero", () => {
 
     rerender(
       <Hero
+        backgroundColor="#654321"
+        foregroundColor="#123456"
         heroType={HeroTypes.Secondary}
         heading={
           <Heading
@@ -272,10 +279,9 @@ describe("Hero", () => {
             text="Hero Secondary"
           />
         }
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
         subHeaderText={subHeaderText}
-        imageComponent={imageComponent}
-        foregroundColor="#123456"
-        backgroundColor="#654321"
       />
     );
 
@@ -298,19 +304,20 @@ describe("Hero", () => {
       <Hero heroType={HeroTypes.Primary} heading={heading} />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: it is recommended to use the "backgroundImageSrc" prop for PRIMARY hero.`
+      `Hero: it is recommended to use the "backgroundImageSrc" prop for the PRIMARY Hero.`
     );
 
     rerender(
       <Hero
+        backgroundImageSrc="https://placeimg.com/1600/800/arch"
         heroType={HeroTypes.Primary}
         heading={heading}
-        backgroundImageSrc="https://placeimg.com/1600/800/arch"
-        imageComponent={imageComponent}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: the "image" prop has been passed, but PRIMARY hero will not use it.`
+      `Hero: the "imageSrc" and "imageAlt" props have been passed, but the "PRIMARY" Hero will not use it.`
     );
   });
 
@@ -327,40 +334,43 @@ describe("Hero", () => {
       <Hero
         heroType={HeroTypes.Secondary}
         heading={heading}
-        subHeaderText={subHeaderText}
-        imageComponent={imageComponent}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
         locationDetails={<>Some location details.</>}
+        subHeaderText={subHeaderText}
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: Please provide "locationDetails" only to PRIMARY hero.`
+      `Hero: Please provide "locationDetails" only to PRIMARY hero.`
     );
 
     rerender(
       <Hero
-        heroType={HeroTypes.Secondary}
-        heading={heading}
-        subHeaderText={subHeaderText}
-        imageComponent={imageComponent}
         backgroundImageSrc="https://placeimg.com/1600/800/arch"
+        heroType={HeroTypes.Secondary}
+        heading={heading}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
+        subHeaderText={subHeaderText}
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: the "backgroundImageSrc" prop has been passed, but SECONDARY hero will not use it.`
+      `Hero: the "backgroundImageSrc" prop has been passed, but SECONDARY hero will not use it.`
     );
 
     rerender(
       <Hero
+        backgroundColor="#654321"
+        foregroundColor="#123456"
         heroType={HeroTypes.Secondary}
         heading={heading}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
         subHeaderText={subHeaderText}
-        imageComponent={imageComponent}
-        foregroundColor="#123456"
-        backgroundColor="#654321"
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: the "foregroundColor" and/or "backgroundColor" props have been passed, but SECONDARY Hero will not use them.`
+      `Hero: the "foregroundColor" and/or "backgroundColor" props have been passed, but SECONDARY Hero will not use them.`
     );
   });
 
@@ -382,31 +392,32 @@ describe("Hero", () => {
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: Please provide "locationDetails" only to PRIMARY hero.`
+      `Hero: Please provide "locationDetails" only to PRIMARY hero.`
     );
 
     rerender(
       <Hero
         heroType={HeroTypes.Tertiary}
         heading={heading}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
         subHeaderText={otherSubHeaderText}
-        imageComponent={imageComponent}
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: TERTIARY hero will not use any of the image props.`
+      `Hero: TERTIARY hero will not use any of the image props.`
     );
 
     rerender(
       <Hero
-        heroType={HeroTypes.Tertiary}
-        heading={heading}
-        subHeaderText={otherSubHeaderText}
         backgroundImageSrc="https://placeimg.com/1600/800/arch"
+        heroType={HeroTypes.Tertiary}
+        heading={heading}
+        subHeaderText={otherSubHeaderText}
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: TERTIARY hero will not use any of the image props.`
+      `Hero: TERTIARY hero will not use any of the image props.`
     );
   });
 
@@ -421,16 +432,31 @@ describe("Hero", () => {
     );
     const { rerender } = render(
       <Hero
+        backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
         heroType={HeroTypes.Campaign}
         heading={heading}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
+        locationDetails={<>Some location details.</>}
         subHeaderText={otherSubHeaderText}
-        backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
-        imageComponent={imageComponent}
+      />
+    );
+    expect(warn).toHaveBeenCalledWith(
+      `Hero: Please provide "locationDetails" only to PRIMARY hero.`
+    );
+
+    rerender(
+      <Hero
+        heroType={HeroTypes.Campaign}
+        heading={heading}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
+        subHeaderText={otherSubHeaderText}
         locationDetails={<>Some location details.</>}
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: Please provide "locationDetails" only to PRIMARY hero.`
+      `Hero: it is recommended to use both "backgroundImageSrc" and "image" props for CAMPAIGN hero.`
     );
 
     rerender(
@@ -438,25 +464,12 @@ describe("Hero", () => {
         heroType={HeroTypes.Campaign}
         heading={heading}
         subHeaderText={otherSubHeaderText}
-        imageComponent={imageComponent}
-        locationDetails={<>Some location details.</>}
-      />
-    );
-    expect(warn).toHaveBeenCalledWith(
-      `Warning: it is recommended to use both "backgroundImageSrc" and "image" props for CAMPAIGN hero.`
-    );
-
-    rerender(
-      <Hero
-        heroType={HeroTypes.Campaign}
-        heading={heading}
-        subHeaderText={otherSubHeaderText}
         backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
         locationDetails={<>Some location details.</>}
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: it is recommended to use both "backgroundImageSrc" and "image" props for CAMPAIGN hero.`
+      `Hero: it is recommended to use both "backgroundImageSrc" and "image" props for CAMPAIGN hero.`
     );
   });
 
@@ -466,24 +479,26 @@ describe("Hero", () => {
       <Hero
         heroType={HeroTypes.FiftyFifty}
         subHeaderText={otherSubHeaderText}
-        imageComponent={imageComponent}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
         locationDetails={<>Some location details.</>}
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: Please provide "locationDetails" only to PRIMARY hero.`
+      `Hero: Please provide "locationDetails" only to PRIMARY hero.`
     );
 
     rerender(
       <Hero
-        heroType={HeroTypes.FiftyFifty}
-        subHeaderText={otherSubHeaderText}
-        imageComponent={imageComponent}
         backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
+        heroType={HeroTypes.FiftyFifty}
+        imageAlt={imageAlt}
+        imageSrc={imageSrc}
+        subHeaderText={otherSubHeaderText}
       />
     );
     expect(warn).toHaveBeenCalledWith(
-      `Warning: the "backgroundImageSrc" prop has been passed, but FIFTYFIFTY hero will not use it.`
+      `Hero: the "backgroundImageSrc" prop has been passed, but FIFTYFIFTY hero will not use it.`
     );
   });
 
@@ -515,8 +530,9 @@ describe("Hero", () => {
               text="Hero Secondary"
             />
           }
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={subHeaderText}
-          imageComponent={imageComponent}
         />
       )
       .toJSON();
@@ -531,8 +547,9 @@ describe("Hero", () => {
               text="Hero Secondary Books and More"
             />
           }
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={subHeaderText}
-          imageComponent={imageComponent}
         />
       )
       .toJSON();
@@ -547,8 +564,9 @@ describe("Hero", () => {
               text="Hero Secondary Locations"
             />
           }
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={subHeaderText}
-          imageComponent={imageComponent}
         />
       )
       .toJSON();
@@ -563,8 +581,9 @@ describe("Hero", () => {
               text="Hero Secondary"
             />
           }
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={subHeaderText}
-          imageComponent={imageComponent}
         />
       )
       .toJSON();
@@ -579,14 +598,16 @@ describe("Hero", () => {
               text="Hero Secondary What's On"
             />
           }
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={subHeaderText}
-          imageComponent={imageComponent}
         />
       )
       .toJSON();
     const campaign = renderer
       .create(
         <Hero
+          backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
           heroType={HeroTypes.Campaign}
           heading={
             <Heading
@@ -595,9 +616,9 @@ describe("Hero", () => {
               text="Hero Campaign"
             />
           }
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={otherSubHeaderText}
-          backgroundImageSrc="https://placeimg.com/2400/800/nature/grayscale"
-          imageComponent={imageComponent}
         />
       )
       .toJSON();
@@ -620,8 +641,9 @@ describe("Hero", () => {
       .create(
         <Hero
           heroType={HeroTypes.FiftyFifty}
+          imageAlt={imageAlt}
+          imageSrc={imageSrc}
           subHeaderText={otherSubHeaderText}
-          imageComponent={imageComponent}
         />
       )
       .toJSON();